requirements.txt
absl-py==0.2.2
albumentations==0.3.1
astor==0.6.2
attrs==19.1.0
backcall==0.1.0
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2019.3.9
chardet==3.0.4
comet-git-pure==0.19.11
configobj==5.0.6
cycler==0.10.0
decorator==4.1.2
dicom==0.9.9.post1
easydict==1.7
enum34==1.1.6
everett==1.0.2
gast==0.2.0
gitdb2==2.0.5
GitPython==2.1.11
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
idna==2.8
imageio==2.8.0
imgaug==0.2.6
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jsonschema==3.0.1
Keras==2.1.6
Keras-Applications==1.0.2
keras-contrib==2.0.8
Keras-Preprocessing==1.0.1
Markdown==2.6.9
matplotlib==2.0.2
netifaces==0.10.9
networkx==1.11
nibabel==2.3.0
numexpr==2.6.2
numpy==1.14.4
nvidia-ml-py3==7.352.0
olefile==0.44
opencv-python-headless==4.1.0.25
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pydicom==1.2.2
Pygments==2.2.0
pyparsing==2.2.0
pyrsistent==0.14.11
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
requests==2.21.0
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
six==1.12.0
smmap2==2.0.5
tensorboard==1.7.0
tensorflow==1.3.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.55.0
Werkzeug==0.12.2
wurlitzer==1.0.2
